Добавлена проверка на нул при детаче. Добавлена обработка на попытку убрать галочки в обоих тоглах. Немного изменил формулы расчета силы противника, чтоб сила постоянно росла, а не падала при развитии игрока. Добавил округление в большую сторону, т.к. сила иногда скакала то в + то в -.

+++ Assets/Scripts/Interfaces/IFightState.cs
@@ -1,4 +1,4 @@
 ﻿public interface IFightState
 {
-    int GetPower(int playerPower, int playerHealth, int playerOverallMoney, int crimeLevelModifer);
+    int GetPower(int playerPower, int playerHealth, int playerOverallMoney, float crimeLevelModifer);
 }
